Support queries through select_related with no None value defined
When Django joins in a table with a EnumField that does not exist,
all the table columns are NULL. The SQL processing then runs the
converters (from_db_value et.al), and the convertion to None for the
ForeignKey field is done later in the manager.

Due to this behaviour we need to return None in our from_db_value method
even if we have no None value in our enum. The defence against null
values is then entirely up to the database to enforce.

Not ideal, but its the only way to get reasonable performance when
querying these tables.