Fix django 3 deprecation warning
According to the [django
2.0](https://docs.djangoproject.com/en/2.0/releases/2.0/#context-argument-of-field-from-db-value-and-expression-convert-value)
docs the `context` argument of `from_db_value` has been deprecated and
will eventually be removed in django 3.

Since the implementation was not using it or `expression` and
`connection` I made them to be swalloed by the args/kwargs so that
it's backwards compatible and hopefully future-proof.